# internship-command-center/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Opportunity {
  id: number;
  company: string;
  role: string;
  location: string;
  mode: string;
  stipend: string;
  stipendMin?: number;
  eligibility: string;
  skills: string[];
  deadline: string;
  applyUrl: string;
  source: string;
  sourceUrl?: string;
  /** @nullable */
  hrEmail?: string | null;
  /** @nullable */
  recruiterLinkedin?: string | null;
  fitScore: number;
  salaryScore: number;
  growthScore: number;
  brandScore: number;
  learningScore: number;
  verified: boolean;
  saved: boolean;
  lastChecked?: string;
  /** @nullable */
  note?: string | null;
}

export interface OpportunityUpdate {
  saved?: boolean;
}

export interface Application {
  id: number;
  opportunityId: number;
  company: string;
  role: string;
  stage: string;
  appliedOn: string;
  nextAction: string;
  /** @nullable */
  notes?: string | null;
}

export interface ApplicationInput {
  opportunityId: number;
  company: string;
  role: string;
  stage: string;
  nextAction: string;
  notes?: string;
}

export interface ApplicationUpdate {
  stage?: string;
  nextAction?: string;
  notes?: string;
}

export interface Profile {
  name: string;
  degree: string;
  year: string;
  country: string;
  locations: string[];
  skills: string[];
  interests: string[];
  target: string;
}

export interface ProfileUpdate {
  name?: string;
  degree?: string;
  year?: string;
  locations?: string[];
  skills?: string[];
  interests?: string[];
  target?: string;
}

export interface DashboardSummary {
  totalOpportunities: number;
  verifiedCount: number;
  topMatch: number;
  applications: number;
  responseRate: number;
  daysLeft: number;
  recommendedToday: Opportunity[];
}

export type SearchParameter = string;

export type LocationParameter = string;

export type ModeParameter = string;

export type FocusParameter = string;

export type MinStipendParameter = number;

export type SortParameter = string;

export type ListOpportunitiesParams = {
search?: SearchParameter;
location?: LocationParameter;
mode?: ModeParameter;
focus?: FocusParameter;
minStipend?: MinStipendParameter;
sort?: SortParameter;
};