# internship-command-center/lib/api-zod/src/generated/types/application.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Application {
  id: number;
  opportunityId: number;
  company: string;
  role: string;
  stage: string;
  appliedOn: string;
  nextAction: string;
  /** @nullable */
  notes?: string | null;
}