# internship-command-center/lib/api-zod/src/generated/types/applicationInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ApplicationInput {
  opportunityId: number;
  company: string;
  role: string;
  stage: string;
  nextAction: string;
  notes?: string;
}