# internship-command-center/lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Opportunity } from './opportunity';

export interface DashboardSummary {
  totalOpportunities: number;
  verifiedCount: number;
  topMatch: number;
  applications: number;
  responseRate: number;
  daysLeft: number;
  recommendedToday: Opportunity[];
}