# internship-command-center/lib/api-zod/src/generated/types/opportunity.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Opportunity {
  id: number;
  company: string;
  role: string;
  location: string;
  mode: string;
  stipend: string;
  stipendMin?: number;
  eligibility: string;
  skills: string[];
  deadline: string;
  applyUrl: string;
  source: string;
  sourceUrl?: string;
  /** @nullable */
  hrEmail?: string | null;
  /** @nullable */
  recruiterLinkedin?: string | null;
  fitScore: number;
  salaryScore: number;
  growthScore: number;
  brandScore: number;
  learningScore: number;
  verified: boolean;
  saved: boolean;
  lastChecked?: string;
  /** @nullable */
  note?: string | null;
}