# internship-command-center/lib/api-zod/src/generated/types/profile.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Profile {
  name: string;
  degree: string;
  year: string;
  country: string;
  locations: string[];
  skills: string[];
  interests: string[];
  target: string;
}